GH-516 Make DataNeedsManagementController use correct HTTP Verbs

+++ core/src/main/java/energy/eddie/core/dataneeds/DataNeedsManagementController.java
@@ -2,18 +2,21 @@ package energy.eddie.core.dataneeds;
 
 import energy.eddie.api.agnostic.DataNeed;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
+import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.lang.NonNull;
 import org.springframework.web.bind.annotation.*;
 
 import java.net.URI;
 
+import static energy.eddie.core.dataneeds.DataNeedsManagementController.URL_PREFIX;
+
 /**
  * REST controller for managing data needs that are stored in the application database. This provides
  * common CRUD operations for data needs as documented in the APIS.md file.
  */
 @RestController
-@RequestMapping("/management/data-needs")
+@RequestMapping(path = URL_PREFIX, produces = MediaType.APPLICATION_JSON_VALUE)
 @ConditionalOnProperty(value = "eddie.data-needs-config.data-need-source", havingValue = "DATABASE")
 public class DataNeedsManagementController {
     public static final String URL_PREFIX = "/management/data-needs";
@@ -24,11 +27,11 @@ public class DataNeedsManagementController {
         this.dataNeedsDbRepository = dataNeedsDbRepository;
     }
 
-    @RequestMapping(method = {RequestMethod.POST, RequestMethod.PUT})
+    @PostMapping(produces = MediaType.TEXT_PLAIN_VALUE, consumes = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<String> createDataNeed(@NonNull @RequestBody DataNeedEntity newDataNeed) {
         final var id = newDataNeed.id();
         if (dataNeedsDbRepository.existsById(id)) {
-            return ResponseEntity.badRequest().body("data need with id " + id + " already exists");
+            return ResponseEntity.status(HttpStatus.CONFLICT).body("data need with id " + id + " already exists");
         }
         dataNeedsDbRepository.save(newDataNeed);
         return ResponseEntity.created(URI.create(URL_PREFIX + "/" + id)).build();
@@ -41,11 +44,12 @@ public class DataNeedsManagementController {
 
     @GetMapping("/{id}")
     public ResponseEntity<DataNeed> getDataNeed(@PathVariable String id) {
-        final var dataNeed = dataNeedsDbRepository.findById(id);
-        return dataNeed.isEmpty() ? ResponseEntity.notFound().build() : ResponseEntity.ok(dataNeed.get());
+        return dataNeedsDbRepository.findById(id)
+                .<ResponseEntity<DataNeed>>map(ResponseEntity::ok)
+                .orElseGet(() -> ResponseEntity.notFound().build());
     }
 
-    @PostMapping("/{id}")
+    @PutMapping(path = "/{id}", produces = MediaType.TEXT_PLAIN_VALUE, consumes = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<String> updateDataNeed(@PathVariable String id, @RequestBody DataNeedEntity dataNeed) {
         if (!dataNeed.id().equals(id)) {
             return ResponseEntity.badRequest().body("data need id in url does not match data need id in body");
@@ -58,8 +62,11 @@ public class DataNeedsManagementController {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Object> deleteDataNeed(@PathVariable String id) {
+    public ResponseEntity<Void> deleteDataNeed(@PathVariable String id) {
+        if (!dataNeedsDbRepository.existsById(id)) {
+            return ResponseEntity.notFound().build();
+        }
         dataNeedsDbRepository.deleteById(id);
-        return ResponseEntity.ok().build();
+        return ResponseEntity.noContent().build();
     }
 }
